requirements.txt
sgp4>=2.22
numpy>=1.24
scipy>=1.10
astropy>=5.3
pyvista>=0.42
pyvistaqt>=0.11
PyQt6>=6.5
vtk>=9.2
requests>=2.31
Pillow>=10.0
matplotlib>=3.7
qtpy>=2.4
